refactor(spec): split strict portability checks from lossy conversion
Extract the PortableSpec walk/isPortable/toPortable path and the explicit
toPortableLossy tooling path into focused modules behind the portability
facade. Co-locate tests. Public portability surface is unchanged.
