+++ .changeset/spec-portability-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Split strict portability checks from lossy tooling conversion, with co-located tests. Public isPortable/toPortable/toPortableLossy surface is unchanged.
